pyproject.toml
[tool.poetry]
name = "gazprominform_testtask"
version = "0.1.0"
description = ""
authors = ["george <klasd4000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
requests-html = "^0.10.0"
pytest-django = "^4.4.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
